+++ crates/rgpui/src/system/svg_renderer.rs
@@ -303,12 +303,12 @@ fn fix_generic_font_families(db: &mut usvg::fontdb::Database) {
 
 #[cfg(test)]
 mod tests {
-    use crate::*;
+    use super::*;
     use usvg::fontdb::{Database, Family, Query};
 
     const IBM_PLEX_REGULAR: &[u8] =
-        include_bytes!("../../../assets/fonts/ibm-plex-sans/IBMPlexSans-Regular.ttf");
-    const LILEX_REGULAR: &[u8] = include_bytes!("../../../assets/fonts/lilex/Lilex-Regular.ttf");
+        include_bytes!("../../../../assets/fonts/ibm-plex-sans/IBMPlexSans-Regular.ttf");
+        const LILEX_REGULAR: &[u8] = include_bytes!("../../../../assets/fonts/lilex/Lilex-Regular.ttf");
 
     fn db_with_bundled_fonts() -> Database {
         let mut db = Database::new();